Adds Function that Checks for Winner
Now that there is a function for checking if there is a winner, we are
able to declare who won the game. Once this declared it sets the
gameOver to true, which prevents any further moves and prints the
winner to the screen.

On reset, this is cleared and gameOver is reverted to false.